+++ src/main/java/webserver/request/support/StaticDirectory.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request.support;
 
 import org.springframework.lang.Nullable;
